pyproject.toml
[tool.poetry]
name = "betsparser"
version = "0.1.0"
description = ""
authors = ["Egor <kogayegor@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
selenium = "^4.18.1"
requests = "^2.31.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
